Move all this AuthToken stuff into its own module
