+++ packages/liveblocks-client/src/types/index.ts
@@ -194,21 +194,6 @@ export type Client = {
   leave(roomId: string): void;
 };
 
-export type AppOnlyAuthToken = {
-  appId: string;
-  roomId?: never;
-};
-
-export type RoomAuthToken = {
-  appId: string;
-  roomId: string;
-  actor: number;
-  id?: string;
-  info?: Json;
-};
-
-export type AuthToken = AppOnlyAuthToken | RoomAuthToken;
-
 /**
  * Represents all the other users connected in the room. Treated as immutable.
  */
